raspberry/requirements.txt
requests>=2.31.0
# Raspberry Pi OS에서 GPIO 모드 사용 시 설치
RPi.GPIO>=0.7.1
